refactor(frontend): Extract key control logic - Extract key control logic into a custom hook
