+++ frontend/src/components/Viewer/DualView.js
@@ -1,7 +1,8 @@
 import React, { useRef, useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
-
 import io from 'socket.io-client';
+
+import { useKeyControl } from './UseKeyControl';
 
 import DualViewHeader from './DualViewHeader';
 import CanvasContainer from './CanvasContainer';
@@ -108,25 +109,7 @@ const DualView = () => {
     };
   }, []);
 
-  useEffect(() => {
-    const keyEventHandler = (event) => {
-      const currentTime = new Date().getTime();
-      if (currentTime - lastKeyPressedTime.current > 30) {
-        lastKeyPressedTime.current = currentTime;
-        if (socketRef.current) {
-          socketRef.current.emit('key_control', { key: event.key, step: step });
-        }
-      } else {
-        console.log('Too many requests!');
-      }
-    };
-
-    window.addEventListener('keypress', keyEventHandler, false);
-
-    return () => {
-      window.removeEventListener('keypress', keyEventHandler, false);
-    };
-  }, [step]);
+  useKeyControl(step, lastKeyPressedTime, socketRef);
 
   const handleResetClick = () => {
     setStep(initStepValue);
