+++ frontend/src/components/Viewer/UseKeyControl.js
@@ -1,0 +1,23 @@
+import { useEffect } from 'react';
+
+export const useKeyControl = (step, lastKeyPressedTime, socketRef) => {
+  useEffect(() => {
+    const keyEventHandler = (event) => {
+      const currentTime = new Date().getTime();
+      if (currentTime - lastKeyPressedTime.current > 30) {
+        lastKeyPressedTime.current = currentTime;
+        if (socketRef.current) {
+          socketRef.current.emit('key_control', { key: event.key, step: step });
+        }
+      } else {
+        console.log('Too many requests!');
+      }
+    };
+
+    window.addEventListener('keypress', keyEventHandler, false);
+
+    return () => {
+      window.removeEventListener('keypress', keyEventHandler, false);
+    };
+  }, [step]);
+};
